Start using a parser combinator instead
This is 'necessary' to more easily parse the conditional tags, which
add quite a bit more complexity that is straightforward to handle with
regexes.

This also means we can reuse the testing strategy from the DBML gem by
having simple tests for each part of the parser inline as comments.