Integration with Jekyll as a plugin
This uses the Jekyll hooks system to start to automatically provide a
ComicFury-compatible environment. The goal is to allow someone to just
blob down their comic pages and blog posts as per normal Jekyll, and
have the plugin do the hard work of setting up the collections,
providing Liquid with the right variables, and of course rendering the
templates into Liquid in the background.

We're going for a 'convention over configuration' approach.

So far, we have:
- putting ComicFury templates into `_layouts` will get picked up and
  rendered by RageRender
- putting HTML or Markdown into `_comics` will get them treated as
  comic pages with the `comic-page` layout automatically applied
- comics by default expect an identically named image in the `images`
  folder, but this can be overwritten with the `image` key in the HTML
  or Markdown front matter
- comics support custom variables in the `custom` key
- comics support chapters in the `chapter` key