Support the Overview page
The Overview page is one of the options for the front page. (In fact,
the overview page is also always available at '/overview', but isn't
linked from anywhere by default.) It shows the latest comic and the
five most recent blogs. As we already know how to render comics and
blogs, providing fields to the overview page just involves reusing
our existing drops.