Automatically generate comic pages for standalone images.
As well as the convention that comic pages pick up images with the same
basename, we can also do the reverse – make it so that a folder full of
comic images will just automatically create comics. This makes it even
easier to get started, because users can just drop their comic images
into the 'images/' folder and render a site.

Created comics have the same title as the image name, pick up the date
from the mtime of the image (as for static files) and inherit the
author from the one set sitewide (the logged in user by default).

If we also do a few more things to handle all of the cases that break
where there is no config file present, we now have a site theme that
can be run against a directory with no config file at all. All you need
is the images and a Gemfile to get a ComicFury-ish site from them!

This does involve setting the theme in the config if one is not
present, which is probably not very hygienic. But this works for now.