Output archive pages
This has been quite a journey. It turns out that there are tons of
different archive pages:

- a complete list of all comics, grouped by chapter, with 160 per page
- a list of comics per chapter, again paginated if the pages in one
  chapter go above 160
- a chapter overview that just shows all the chapters
- a "main" archive page that shows one of these pages depending on if
  chapters are being used or not

So there has been lots of fun getting generators to create the right
structures for all the different cases, and then making sure all of
the pages end up with the correct variables.

There is a missing "chapter" object, because it turns out that chapters
can have covers and HTML descriptions of their own. So we need some
object to put them into. However, having the default be based on
chapter slugs in comic meta works well, so we can continue to use that
with a generator. We should also probably have more automagic where
images in folders cause chapters to get created, or something.