Refactor to use a new collection for chapters
Chapters now get to be first class objects. This… kind of is better?
It certainly is a bit more consistent but not much code has disappeared.
I suppose the benefit is that users can now set chapter images and
descriptions explicitly now, which wasn't possible before.

The Generator system is creaking a little bit under the weight of so
many dependent things. Generators only have a single 'priority' field
and there are only five values. It's getting hard to express/remember
that generator X should really be called before generator Y. With any
more generators it might be necessary to intervene here and set up
some kind of explicit dependencies.