Apply site variables to pages as well
Previously we didn't actually apply any variables to site pages. But
things still kinda worked. It turns out that when we merge fields into
the payload, we are actually mutating a site-wide hash of data that is
used for every page. This only works most of the time because Jekyll
correctly overrides the variables every time, so stale data is always
removed. But because we weren't overriding data, site-wide variables
were still present for pages but took values of the last post or comic
that was generated. Tbh I think this is a poor design on Jekyll's part
but is probably performance motivated.

Now we definitely set site-wide variables for all documents and pages,
so correctly override them every time.