Support for a 'static' search page
Jekyll won't support a dynamic search page, but for testing purposes
we can render static search page(s) that allow search result designs
to be tested. This actually works for multiple search pages, just by
setting the layout to search and setting a `searchterm` in the front
matter. But for the default page, config can use:

  - scope:
      path: assets/search.html
    values:
      searchterm: ralsei