Escape quotes properly in variable output
ComicFury will automatically HTML-escape quotes and apostrophes when
outputting variables (so that they can be safely included in HTML
attributes). The `rawhtml` function prevents this escaping from
happening. This doesn't happen for string literals – only for variables.

So, we now properly escape apostrophes and quotes when rendering
variables, and the rawhtml function is now implemented.