Support quotes in string literals
Neither Liquid nor ERB could handle having a quote in a string literal
in a ComicFury function call or comparison, because in neither case
did we do any escaping.

The answer for ERB was easy – just escape the quotes. Liquid is not so
simple, because it doesn't support escape sequences. The only way we
could do this was to use {% capture %} syntax to get the literal string
containing quotes into a variable. This perhaps still has an edge case
if a variable contains the literal string "{% endcapture %}" but fuck
that basically.