Correctly escape user-supplied variables
The purpose of variable escaping is to stop rogue characters in user
data from breaking page templates. This commit manually escapes all
of the variables that could contain user-defined data.

Most variables are always escaped, but some are only escaped if they
came from a non-HTML source. Blog content and author notes that are
picked up from files with a .html extension are passed through without
modification.

There is still some work to do: custom variables can be marked as
containing HTML or not, and we don't have a way of representing that
currently.

Closes #7.