Clean up Jekyll payload before rendering documents
Due to the way we integrated with the Jekyll drop system, stale data
from a previous document can hang around if it isn't explicitly
overwritten in the current document. This can happen when a document
of one type is written after a document of another, e.g. if a comic
is written after a blog post, it would end up with the [v:blogtitle]
of the last blog being set.

Now, we explicitly clear the payload of any RageRender-set variables
before we start, so that stale data is cleared out at the start of
every render.

Resolves #2. Resolves #3.